Working Javascript 😧
First time using local storage for javascript to have the shopping cart work right.
Also added an image slider that I made in a previous project.
Added a full HTML input form that has javascript validate the form and show a dialog box to to notify user of error or completion.